Renamed utils.h to Utils.h and added more test glb models

+++ src/AssetManager/Mesh/Index.h
@@ -7,7 +7,7 @@
 import vulkan_hpp;
 #endif
 
-typedef uint16_t meshIndex_t;
+typedef uint32_t meshIndex_t;
 
 // Compile-time helper function
 constexpr VkIndexType getVkIndexType() {
